+++ packages/ag-grid-community/src/clientSideRowModel/clientSideRowModel.ts
@@ -607,14 +607,6 @@ export class ClientSideRowModel extends BeanStub implements IClientSideRowModel,
         return true; // Nothing changed, or only updates with no new rows and no removals
     }
 
-    private beforeRefreshModel(params: RefreshModelParams, groupsChanged: boolean = false): void {
-        this.eventSvc.dispatchEvent({ type: 'beforeRefreshModel', params, groupsChanged });
-
-        if (this.started && params.rowDataUpdated) {
-            this.eventSvc.dispatchEvent({ type: 'rowDataUpdated' });
-        }
-    }
-
     public refreshModel(params: RefreshModelParams): void {
         if (!this.rootNode) {
             return; // Destroyed
@@ -632,7 +624,12 @@ export class ClientSideRowModel extends BeanStub implements IClientSideRowModel,
         // let start: number;
         // console.log('======= start =======');
 
+        const beans = this.beans;
         const changedPath = (params.changedPath ??= this.createChangePath(!params.newData && !!params.rowDataUpdated));
+
+        if (this.started && params.rowDataUpdated) {
+            this.eventSvc.dispatchEvent({ type: 'rowDataUpdated' });
+        }
 
         if (
             !this.started ||
@@ -645,17 +642,19 @@ export class ClientSideRowModel extends BeanStub implements IClientSideRowModel,
 
         this.isRefreshingModel = true;
 
-        this.beans.masterDetailSvc?.refreshModel(params);
+        beans.masterDetailSvc?.refreshModel(params);
 
-        if (params.step !== 'group') {
-            this.beforeRefreshModel(params);
+        if (params.rowDataUpdated && params.step !== 'group') {
+            beans.colFilter?.refreshModel();
         }
 
         /* eslint-disable no-fallthrough */
         switch (params.step) {
             case 'group': {
                 const groupingChanged = this.doRowGrouping(params);
-                this.beforeRefreshModel(params, groupingChanged); // Do this after grouping, so the parent field is correct
+                if (groupingChanged || params.rowDataUpdated) {
+                    beans.colFilter?.refreshModel();
+                }
                 if (params.step === 'group' && this.rowNodesCountReady) {
                     this.rowCountReady = true; // only if row data has been set
                     this.eventSvc.dispatchEventOnce({ type: 'rowCountReady' });
