+++ packages/ag-grid-enterprise/src/rowHierarchy/rowHierarchyUtils.ts
@@ -5,6 +5,7 @@ import type {
     Column,
     GridOptionsService,
     GroupingApproach,
+    IRowNode,
     RowNode,
     StageExecuteParams,
 } from 'ag-grid-community';
@@ -80,3 +81,22 @@ export function _getGroupValue(
     }
     return null;
 }
+
+export const _getRowDefaultExpanded = (beans: BeanCollection, rowNode: IRowNode, level: number): boolean => {
+    const gos = beans.gos;
+    const isGroupOpenByDefault = gos.get('isGroupOpenByDefault');
+    if (!isGroupOpenByDefault) {
+        const groupDefaultExpanded = gos.get('groupDefaultExpanded');
+        return groupDefaultExpanded === -1 || level < groupDefaultExpanded;
+    }
+    const params = {
+        api: beans.gridApi,
+        context: beans.gridOptions.context,
+        rowNode,
+        field: rowNode.field!,
+        key: rowNode.key!,
+        level,
+        rowGroupColumn: rowNode.rowGroupColumn!,
+    };
+    return isGroupOpenByDefault(params) == true;
+};
