+++ packages/ag-grid-enterprise/src/treeData/treeGroupStrategy.ts
@@ -3,6 +3,7 @@ import { RowNode, _ROW_ID_PREFIX_ROW_GROUP, _removeFromArray } from 'ag-grid-com
 import { BeanStub, _EmptyArray, _warn } from 'ag-grid-community';
 
 import { setRowNodeGroup } from '../rowGrouping/rowGroupingUtils';
+import { _getRowDefaultExpanded } from '../rowHierarchy/rowHierarchyUtils';
 import type { GroupingRowNode, IRowGroupingStrategy } from '../rowHierarchy/rowHierarchyUtils';
 import type { DataFieldGetter } from './fieldAccess';
 import { makeFieldPathGetter } from './fieldAccess';
@@ -318,7 +319,7 @@ export class TreeGroupStrategy<TData = any> extends BeanStub implements IRowGrou
             }
         } else if ((flags & FLAG_EXPANDED_INITIALIZED) === 0) {
             row.treeNodeFlags |= FLAG_EXPANDED_INITIALIZED;
-            row.expanded = this.getRowDefaultExpanded(row, level); // Initialize the expanded state
+            row.expanded = _getRowDefaultExpanded(this.beans, row, level); // Initialize the expanded state
         }
 
         if (collapsed && row.rowIndex !== null) {
@@ -377,25 +378,6 @@ export class TreeGroupStrategy<TData = any> extends BeanStub implements IRowGrou
             }
         }
         return trulyChanged;
-    }
-
-    private getRowDefaultExpanded(rowNode: GroupingRowNode<TData>, level: number): boolean {
-        const { gos, beans } = this;
-        const isGroupOpenByDefault = gos.get('isGroupOpenByDefault');
-        if (!isGroupOpenByDefault) {
-            const groupDefaultExpanded = gos.get('groupDefaultExpanded');
-            return groupDefaultExpanded === -1 || level < groupDefaultExpanded;
-        }
-        const params = {
-            api: beans.gridApi,
-            context: beans.gridOptions.context,
-            rowNode,
-            field: rowNode.field!,
-            key: rowNode.key!,
-            level,
-            rowGroupColumn: rowNode.rowGroupColumn!,
-        };
-        return isGroupOpenByDefault(params) == true;
     }
 
     /** Handle cycles in a tree. Is not optimal for performance but this is an edge case that shouldn't happen as is a warning. */
